requirements.txt
discord.py[speed]>=2.6.2
asyncpg>=0.12.0
redis>=6.4.0
typing_extensions>=4,<5
click
lru_dict
